Refactor cursor movement into a dedicated function
Soon the values at the bottom which show information about the currently
selected byte will be cached. To do this we need to update the text each
time the cursor moves. This helps keep all that in one place.
Note that this causes some regressions, the view no longer scrolls when
hitting the top or bottom rows. It was buggy anyway :)