Jai rewrite.
This commit replaces the C++ codebase with a Jai version. The raylib
dependency meant we relied on raylib's way of doing things. It did give
and easy start but the event pumped frame loop meant that CPU and GPU
usage were unnecessarily high for an application that rarely repaints.

A quick re-write in Jai solved the CPU and GPU issue and gave me chance
try out and hack at the realtively new GUI library that Jai provides.