Build as a GUI application.
This commit makes the metaprogram link the executable as a WINDOWS
application, instead of CONSOLE.

It meant I had to create my own wWinMain and pipe the command line args
into jai's expected formats manually as this is not currently supported.

This introudces the problem of the executable not properly terminating.